split + clean up unrealsdk.cpp
